+++ packages/coding-agent/CHANGELOG.md
@@ -206,6 +206,9 @@
 ### Changed
 
 - Load persisted Agent Hub subagents asynchronously to avoid blocking the TUI on synchronous directory walks ([#4239](https://github.com/can1357/oh-my-pi/issues/4239))
+### Fixed
+
+- Cache failed `!command` config resolutions and time out extension dynamic model fetches after 15s ([#4237](https://github.com/can1357/oh-my-pi/issues/4237))
 
 
 ## [16.3.0] - 2026-07-02